+++ tweetit.js
@@ -1,6 +1,7 @@
 var config = require('./config.json'),
   childProcess = require('child_process'),
   fs = require('fs'),
+  gm = require('gm'),
   twitterAPI = require('node-twitter-api'),
   twitter = new twitterAPI({
     consumerKey: config.twitterConsumerKey,
@@ -34,35 +35,52 @@ function convertToGif(fileParts, word) {
   filename = fileParts[0];
 
   // Create gif
-  childProcess.execSync('convert scrap/' + fileParts.join('.') + ' -page +2+0 -background white -flatten scrap/' + filename + '-1.'+ extension);
-  childProcess.execSync('convert scrap/' + fileParts.join('.') + ' -page +0+2 -background white -flatten scrap/' + filename + '-2.' + extension);
-  childProcess.execSync('convert scrap/' + fileParts.join('.') + ' -page -2+0 -background white -flatten scrap/' + filename + '-3.' + extension);
-  childProcess.execSync('convert scrap/' + fileParts.join('.') + ' -page +0-2 -background white -flatten scrap/' + filename + '-4.' + extension);
-  childProcess.execSync('convert -delay 2 `seq -f scrap/'+filename+'-%01g.'+extension+' 1 1 4` -coalesce final/'+ filename + '.gif');
-  childProcess.execSync('convert final/' + filename + '.gif -stroke "#000" -strokewidth 1 -font ' + config.font + ' -pointsize 24 -gravity south -fill white -annotate +0+10 "[' + word + ' intensifies]" final/' + filename + '.gif');
+  var counter = 0,
+    originalImage = fs.readFileSync('scrap/' + fileParts.join('.')),
+    offsets = ['+2+0', '+0+2', '-2+0', '+0-2'];
 
-  // Cleanup
-  childProcess.exec('rm scrap/' + filename + '*');
-
-  twitter.statuses("update_with_media", {
-      status: "[" + word + " intensifies]",
-      media: [
-      "final/"+filename+".gif",
-      ]
-    },
-    config.twitterAccessToken,
-    config.twitterAccessSecret,
-    function(error, data, response) {
-      if (error) {
-        // something went wrong
-        console.log(error);
-        process.exit();
-      } else {
-        // data contains the data sent by twitter
-        console.log(data);
-        process.exit();
+  gm(originalImage, fileParts.join('.'))
+    .options({imageMagick: true})
+    .size(function(err, val) {
+      for (var i = 0; i <= 3; i++) {
+        gm(originalImage)
+          .options({imageMagick: true})
+          .page(val.width, val.height, offsets[i])
+          .flatten()
+          .write('scrap/' + filename + '-' + i +'.' + extension, function() {
+            counter++;
+            if (counter === 4) {
+              childProcess.execSync('convert -delay 2 `seq -f scrap/'+filename+'-%01g.'+extension+' 0 1 3` -coalesce final/'+ filename + '.gif');
+              childProcess.execSync('convert final/' + filename + '.gif -stroke "#000" -strokewidth 1 -font '+ config.font +' -pointsize 24 -gravity south -fill white -annotate +0+10 "[' + word + ' intensifies]" final/' + filename + '.gif');
+              childProcess.exec('rm scrap/' + filename + '*');
+              gm('final/'+filename+'.gif')
+                .options({imageMagick: true})
+                .crop(val.width - 4, val.height - 4, 2, 2)
+                .write('final/' + filename + '.gif', function() {
+                  twitter.statuses("update_with_media", {
+                      status: "[" + word + " intensifies]",
+                      media: [
+                      "final/"+filename+".gif",
+                      ]
+                    },
+                    config.twitterAccessToken,
+                    config.twitterAccessSecret,
+                    function(error, data, response) {
+                      if (error) {
+                        // something went wrong
+                        console.log(error);
+                        process.exit();
+                      } else {
+                        // data contains the data sent by twitter
+                        console.log(data);
+                        process.exit();
+                      }
+                  });
+                });
+            }
+          })
       }
-  });
+    });
 }
 
 function getRandomWord(callback) {
